made changes per review
curly braces
var -> let
num_users -> overflow_id as the counter really only serves as a way of ensuring every user has an icon. we don't decrement to protect from users appearing with the same id in the icon.

we could just use the socket id for a truly unique icon but that's less readable