move firstlogin check to ensureauthenticated
idea is that now e4p doesnt need to use our debug account to do the initial login and setup their account. in the old method, our debug account would persist in the database. this way, we don't need to give them a database with a dummy user initially, just a db with the right tables setup. when they first login, theyll create a user and that will be the only user.